lots of implementations...
1. add get pndg sparsity option in creating functionspace. the get_pndg_sparsity function is in sparsity.py

2. add pndg prolongator to prolongate pressure to the same space as velocity, we that we can write them in a single vtu file. the prolongator can be computed by calling function_space.get_pndg_prolongator.

3. add fsi_output.py, make tweaks to output.py. we will output to two filed, one for fluid vel/pressure, another for displacement (both structure and mesh displacement).

4. tweak bugs in slicing_x_i

5. project mesh displacement to make it C1 continuous and strongly enforce 0 boundary condition on \partial\Omega_f.

+++ z05-ns/shape_function.py
@@ -1732,7 +1732,7 @@ def sdet_snlx_3d(snlx, x_loc, sweight, nloc, sngi, sn=None, real_snlx=None):
             'fjng,bin,fn,fjk->bfigk',
             snlx,
             x_loc,
-            sn.sum(-1) != 0,
+            (sn.sum(-1) != 0).to(torch.float64),
             drst_duv,
         )
         ddu_x_ddv = torch.linalg.cross(
